+++ stats/opentelemetry/grpc_trace_bin_propagator.go
@@ -16,7 +16,7 @@
  *
  */
 
-package experimental
+package opentelemetry
 
 import (
 	"context"
