+++ stats/opentelemetry/grpc_trace_bin_propagator_test.go
@@ -16,7 +16,7 @@
  *
  */
 
-package experimental
+package opentelemetry
 
 import (
 	"context"
@@ -24,18 +24,9 @@ import (
 
 	"github.com/google/go-cmp/cmp"
 	oteltrace "go.opentelemetry.io/otel/trace"
-	"google.golang.org/grpc/internal/grpctest"
 	"google.golang.org/grpc/metadata"
 	itracing "google.golang.org/grpc/stats/opentelemetry/internal/tracing"
 )
-
-type s struct {
-	grpctest.Tester
-}
-
-func Test(t *testing.T) {
-	grpctest.RunSubTests(t, s{})
-}
 
 var validSpanContext = oteltrace.SpanContext{}.WithTraceID(
 	oteltrace.TraceID{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16}).WithSpanID(
